requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
sqlalchemy>=2.0.20
python-dotenv>=1.0.0
tabulate>=0.9.0
langchain-groq>=0.3.0
